Convert all the regions and divs
I don't properly understand why, but I noticed that during conversion of a test file only 1 div was being converted and not all the regions were. Styles and p elements were _apparently_ fine.

This commit makes all the regions and divs convert correctly, but the `_fix_fontsize()` call that creates `fontSize` and `lineHeight` styles is not currently being run.

It is possible that the buggy behaviour was introduced with Python 3.7.5, since this seemed to work in 3.7.4 previously.