Tidy nodes so we can call their functionality (#45)
* Make Denester behave like a proper node

* Takes a node id and a sequence id and uses the sequence id in the output
* pep8 tick
* checks for seeing a document already and not reprocessing if so, and checks for document being a subtitle document not a live message
* `process_document()` does `emit_data()` properly and does not return the document.
* main work is delegated to a separate `denest()` method that does return the document.
* tests updated to use the `denest()` method.

The Denester still has a lot of class methods that are not bound to either `self` or `cls` but work anyway. They're really like scoped public functions that work independent of any particular object. I left that as is.

* Factor out document conversion

Create a separate callable function for converting documents and use that in `process_document` so that it is also available in other contexts, where full node carriage functionality is not needed.

* Tidy the EBU-TT-D Encoder

* pep8 tick
* deal with unexpected kinds of documents better

* Extract document conversion

Allows the EBUTTDEncoder's document conversion function to be called outside of `process_document` to allow plumbing to be done manually.

* Address review comment

Decorate static methods with `@staticmethod`

* Fix object types for EBU-TT-D head and span

We accidentally removed code to call `_SetSupersedingClass()` from the `d_head_type` and `d_span_type`, the result of which was that when serialising to XML they got an unwanted `xsi:type` attribute. Fix that!

* Fix default document handling

When the `EBUTTDEncoder` sets up a default document, it should use those parts that are important as defaults for the output documents. Do this for `lang` and `implicit_ns` and make the `implicit_ns` handling more elegant in `EBUTTDDocument` by using proper setters and getters.

* Fix namespace of tt metadata elements

Elements defined within other elements were getting global namespace instead of qualified namespace, so correct content was failing validation. Fix that!

* pep8 tick

"Sledgehammer" that also resolves the single out of place `(` review comment "nut"!

* Add an extra delimiter to the documentOriginatingSystem metadata

Looked a bit silly before.