Fix missing elements in EBU-TT-D conversion output
Testing with active area calculation revealed that only one `div` element was ever being made. That wasn't ideal!

This refactor explicitly makes the `body`, `div` and `p` elements, and fixes the font style up.

While I was at it, I realised that the function to fix up font styles assumed that there would either always be both `fontSize` and `lineHeight`, or neither. So allow it to work when there is either one or the other or both.

That required a change to autogenerated style ids and the tests that relied on them.

The other thing that was happening was that `fontSize` and `lineHeight` were only ever being specified on styles for `p` and `span` elements.

This change correctly applies `fontSize` and `lineHeight` styles in `%` units for styles set on `body` and `div` elements too.