MSPA 427 Add IMSC attributes, calculate activeArea and fix EBU-TT-D conversion bugs (#46)
* Fix object types for EBU-TT-D head and span

We accidentally removed code to call `_SetSupersedingClass()` from the `d_head_type` and `d_span_type`, the result of which was that when serialising to XML they got an unwanted `xsi:type` attribute. Fix that!

* pep8 tick

"Sledgehammer" that also resolves the single out of place `(` review comment "nut"!

* Add IMSC attributes

Allow documents to include ittp:activeArea and itts:filLineGap.

* Add documentation and include new attributes in documents

* Add option to calculate the active area

* Test calculation of activeArea

* Convert all the regions and divs

I don't properly understand why, but I noticed that during conversion of a test file only 1 div was being converted and not all the regions were. Styles and p elements were _apparently_ fine.

This commit makes all the regions and divs convert correctly, but the `_fix_fontsize()` call that creates `fontSize` and `lineHeight` styles is not currently being run.

It is possible that the buggy behaviour was introduced with Python 3.7.5, since this seemed to work in 3.7.4 previously.

* Make sure styles are computed before processing

* Fix missing elements in EBU-TT-D conversion output

Testing with active area calculation revealed that only one `div` element was ever being made. That wasn't ideal!

This refactor explicitly makes the `body`, `div` and `p` elements, and fixes the font style up.

While I was at it, I realised that the function to fix up font styles assumed that there would either always be both `fontSize` and `lineHeight`, or neither. So allow it to work when there is either one or the other or both.

That required a change to autogenerated style ids and the tests that relied on them.

The other thing that was happening was that `fontSize` and `lineHeight` were only ever being specified on styles for `p` and `span` elements.

This change correctly applies `fontSize` and `lineHeight` styles in `%` units for styles set on `body` and `div` elements too.

* Tests should work without validating input document

This functionality has been moved to the encoder itself.

* Don't print styles

* Test that the right metadata is being added by the encoder

* address review comments

* NO-TICKET fix import for project (#47)

* NO-TICKET testing if import works

* NO-TICKET change import for docs

* Fix documentation indent bug

Removes a warning