Mspa 427 denester and ebuttd converter times fix (#48)
* Fix object types for EBU-TT-D head and span

We accidentally removed code to call `_SetSupersedingClass()` from the `d_head_type` and `d_span_type`, the result of which was that when serialising to XML they got an unwanted `xsi:type` attribute. Fix that!

* pep8 tick

"Sledgehammer" that also resolves the single out of place `(` review comment "nut"!

* Add IMSC attributes

Allow documents to include ittp:activeArea and itts:filLineGap.

* Add documentation and include new attributes in documents

* Add option to calculate the active area

* Test calculation of activeArea

* Convert all the regions and divs

I don't properly understand why, but I noticed that during conversion of a test file only 1 div was being converted and not all the regions were. Styles and p elements were _apparently_ fine.

This commit makes all the regions and divs convert correctly, but the `_fix_fontsize()` call that creates `fontSize` and `lineHeight` styles is not currently being run.

It is possible that the buggy behaviour was introduced with Python 3.7.5, since this seemed to work in 3.7.4 previously.

* Make sure styles are computed before processing

* Fix missing elements in EBU-TT-D conversion output

Testing with active area calculation revealed that only one `div` element was ever being made. That wasn't ideal!

This refactor explicitly makes the `body`, `div` and `p` elements, and fixes the font style up.

While I was at it, I realised that the function to fix up font styles assumed that there would either always be both `fontSize` and `lineHeight`, or neither. So allow it to work when there is either one or the other or both.

That required a change to autogenerated style ids and the tests that relied on them.

The other thing that was happening was that `fontSize` and `lineHeight` were only ever being specified on styles for `p` and `span` elements.

This change correctly applies `fontSize` and `lineHeight` styles in `%` units for styles set on `body` and `div` elements too.

* Tests should work without validating input document

This functionality has been moved to the encoder itself.

* Don't print styles

* Test that the right metadata is being added by the encoder

* address review comments

* NO-TICKET fix import for project (#47)

* NO-TICKET testing if import works

* NO-TICKET change import for docs

* Fix documentation indent bug

Removes a warning

* Denester fix: Only write span begin and end times if they differ from the parent

Stops us from writing unnecessary times into spans.

This turned out to be a bit nasty, because timed elements return a computed time that can depend on their _children_. So we had to work out if the parents are specifying the end time or if it is this element doing it, and then deal with it on that basis.

Some tests updated too.

* Write times to the right elements

If there are times on spans, don't write times to their parent p elements, but if there are no times on spans, then do write times to their parent p elements. As per Tech3380 requirements for begin and end timing attributes on p and span elements.

* Allow indent and newlines to be specified when getting xml

* Address review comment

Remove unnecessary test checks introduced accidentally in 9e962d7

* Make pushed time calculation less brittle

and pass pep8/flake8.

The previous iteration relied on there being an equal number of begin and end times pushed to the dataset and had no plan for what to do if the calling code had not done that properly. It shouldn't have been a problem but was making me uncomfortable for later.

Instead, push pairs of begin and end times as named tuples, and be less brittle about the key used to get the list from the dataset by defining it globally.

There's probably a way to refactor this even more, but this is an improvement over the first iteration, at least.

* Copy copyright across

* Round autogen style horiz and vert to 2 dp

Also use 'n' instead of 'None'. Both changes save a bit of space and are less prone to rounding errors causing odd presentational differences.

* Remove commented out lines

* No longer comment out documentStartOfProgramme

The code now deals with that element without raising an uncaught exception.

* No orphan spans, and copy span children properly

* copy lang in div

* Tidy removed divs better, merge metadata

When removing divs after unnesting, leaves the `body` element in a tidy state with no orphan `div` elements in `orderedContent()`.

Treats `div`s with different metadata as being different, and merges metadata when denesting. This may not always be the best strategy, but it at least preserves the input data somehow.

* Improves tests

* Removes unnecessary regex commenting of `ebuttm:documentStartOfProgramme`, which the library now handles fine.
* Sets up tests to work with the recently updated denester code by adding a metadata key to the attribute data sets where needed.
* Tests that metadata gets merged correctly.
* Tests that `br` elements are preserved.

* Improve documentation

* Add Resequencer to Scripts and their functions and explain about the use of Denester and Resequencer in the context of the EBU-TT-D Encoder.
* Describe the implementation approach of Denester including combining metadata and complications of removing elements in pyxb.

Co-authored-by: Muad Mohamed <muad.mohamed@bbc.co.uk> with help from Daniel Pope <daniel.pope@bbc.co.uk>.