Add BDD test for checking added styles make it through EBU-TT-D
Adds a specific test case known to trigger a failure before, where the style `"S2S6"` was created and added by the Denester but in a way that the EBUTTDEncoder couldn't see.

Also refactors style checks on spans so that any span's style can be tested, and adds the ability to check that an identified style does actually exist.

Make the nested span conversion ebutt1->ebuttd end to end

We didn't have an end to end EBU-TT Part 1 -> EBU-TT-D test before, and this adds one, and tests denesting creation of styles too.