Denester: correct timing type for the document timebase
Reverts `ee5845b`'s temporary workarounds to `testing/bdd/templates/nested_elements_hardcoded.xml` and `testing/bdd/templates/nested_spans_hardcoded.xml`

The reason they failed is because Denester always created a `FullClockTimingType` for any times it modified, regardless of the document's `timeBase`. That worked for `media` but not for `clock`, for which the specification requires that the more constrained `LimitedClockTimingType` is used.

Fix Denester to create the appropriate type of time depending on the document's timebase, and fix the unit tests to provide the document object in the dataset so that it can do this.

Ensures that in every test where conversion to EBU-TT-D occurs after denesting, the denested document is validated, otherwise a SemanticValidationError is (correctly) raised because some necessary computations must be done before the conversion.

Denester: make the right kind of percentage font size type

Super subtle issue that caused downstream problems. Denester was making a style to set font size, and when it made a percentage one, it was using `ebuttdt.percentageFontSizeType`; this was persisting as the computed font size after validation, however when using the computed font size downstream, for example in the `EBUTT3EBUTTDConverter`, it is important to have access to the `.vertical()` method of the derived class `PercentageFontSizeType`: not having that caused test failures. Therefore, make the right type of percentage font size type, with a capital P, and the tests pass.